Move helper functions to helper file
Some of the helper functions were defined in root.rc, others in
helpers.rc. This moves them all into helper.rc.
Additionally, the helpers.rc file will now be sourced at the beginning
of root.rc instead of during the "general" loading part (thus the name
change, this avoids loading the file later). I'm not sure if that is a
good solution, but it works for now.