Feature/windows install (#7)
* Check for make in path before building ble

* Make shell type detection more stable

* Make install script windows compatible

* fixup! Check for make in path before building ble

* Move helper functions to dedicated file

* Only load modules if not on windows

* fixup! Move helper functions to dedicated file

* Use source instead of dot operator

* Move helper functions to helper file

Some of the helper functions were defined in root.rc, others in
helpers.rc. This moves them all into helper.rc.
Additionally, the helpers.rc file will now be sourced at the beginning
of root.rc instead of during the "general" loading part (thus the name
change, this avoids loading the file later). I'm not sure if that is a
good solution, but it works for now.

* Enable ble.sh on windows

Note: Since ble.sh needs to be compiled, it may need MSYS or another
terminal with "make" functionality!

* fixup! Move helper functions to helper file

* Add expected env variables on windows

Ble.sh expects some unix env variables to be set - this takes care of
that if we detect windows

* fixup! Move helper functions to helper file

---------

Co-authored-by: Steffen Peikert <steffen.peikert@gmail.com>