Cleanup of tmux start
Instead of a convoluted start/banner check/passing on tmux start, it is
now handled in the on-session-start script which gets invoked by tmux.
This also makes the startup in bashrc a lot less convoluted :party:

Furthermore, the bashrc now executes tmux rather than treating it as a
dedicated command, which effectively replaces the current shell by tmux.
Just nicer and cleaner, and does not require to manually terminate the
"old" running shell