Feature/tmux before dotfiles (#17)
* TMUX is now run before the dotfiles hook

* Destroy finished tmux sessions

TMUX sessions were kept alive (i.e., detached) if the user simply closed
the terminal session via the "x" button. Now, tmux will unalive all
unattached sessions that are created from an interactive shell. This
does not affect sessions that are purposely detached.

* Enable (stacked) copy-paste

* Ignore dupes in bash history

* Keep tmux sessions alive per default

* Perform autoupdate of ble

Added a function that returns true if a specified intervall of x seconds
has passed since the last check. This can be used for performing timed
checks/updates (e.g., as for ble.sh)

Also ran formatter on the root script and the modules script

* Remove in-tree build of ble

This creates an error everytime we are checking if there is an update to
ble. By moving it to a dedicated out-of-tree directory, no issue arises

* Cleanup of tmux start

Instead of a convoluted start/banner check/passing on tmux start, it is
now handled in the on-session-start script which gets invoked by tmux.
This also makes the startup in bashrc a lot less convoluted :party:

Furthermore, the bashrc now executes tmux rather than treating it as a
dedicated command, which effectively replaces the current shell by tmux.
Just nicer and cleaner, and does not require to manually terminate the
"old" running shell

* fixup! Cleanup of tmux start

* fixup! Cleanup of tmux start